+++ src/dto/audit.js
@@ -11,7 +11,11 @@
  */
 
 import {
-  createAudit, AUDIT_TYPE_BROKEN_BACKLINKS, AUDIT_TYPE_ORGANIC_TRAFFIC, AUDIT_TYPE_ORGANIC_KEYWORDS,
+  createAudit,
+  AUDIT_TYPE_BROKEN_BACKLINKS,
+  AUDIT_TYPE_ORGANIC_TRAFFIC,
+  AUDIT_TYPE_ORGANIC_KEYWORDS,
+  AUDIT_TYPE_CWV,
 } from '@adobe/spacecat-shared-data-access/src/models/audit.js';
 
 /**
@@ -79,7 +83,8 @@ export const AuditDto = {
   toAbbreviatedJSON: (audit) => {
     if (audit.getAuditType() === AUDIT_TYPE_BROKEN_BACKLINKS
       || audit.getAuditType() === AUDIT_TYPE_ORGANIC_KEYWORDS
-      || audit.getAuditType() === AUDIT_TYPE_ORGANIC_TRAFFIC) {
+      || audit.getAuditType() === AUDIT_TYPE_ORGANIC_TRAFFIC
+      || audit.getAuditType() === AUDIT_TYPE_CWV) {
       return AuditDto.toJSON(audit);
     }
     return {